Fix error with DELETE verb
An internal server error was thrown when an attempt was made to delete
an amenity with a non-existent id. This commit fixes the error so that
now a 404 error is thrown.

+++ api/v1/views/amenities.py
@@ -25,6 +25,8 @@ def amenities(amenity_id=None):
         # Retrieve the specific amenity because it exists
         return jsonify(amenity_object.to_dict())
     elif request.method == 'DELETE':
+        if amenity_object is None:
+            abort(404)
         amenity_object.delete()
         storage.save()
         return jsonify({}), 200
